+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 -   `emailAttachmentsEndpoint` to `FormSubmissionEventConfiguration` class
 -   `UploadEmailAttachment` to `FormsClient` class
+-   `WorkflowAttachmentUploadCredentials` to models
+-   `EmailAttachmentData` to models
 
 ## 5.2.0 (2023-11-21)